+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/nodes/gc/CardTableBarrierSet.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2022, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2025, Oracle and/or its affiliates. All rights reserved.
  * Copyright (c) 2019, Red Hat Inc. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -32,7 +32,6 @@ import jdk.graal.compiler.core.common.type.AbstractObjectStamp;
 import jdk.graal.compiler.core.common.type.Stamp;
 import jdk.graal.compiler.debug.Assertions;
 import jdk.graal.compiler.debug.GraalError;
-import jdk.graal.compiler.nodes.FixedWithNextNode;
 import jdk.graal.compiler.nodes.NodeView;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.ValueNode;
@@ -43,8 +42,8 @@ import jdk.graal.compiler.nodes.java.LoweredAtomicReadAndWriteNode;
 import jdk.graal.compiler.nodes.memory.FixedAccessNode;
 import jdk.graal.compiler.nodes.memory.ReadNode;
 import jdk.graal.compiler.nodes.memory.WriteNode;
-import jdk.graal.compiler.nodes.memory.address.AddressNode;
 import jdk.graal.compiler.nodes.memory.address.OffsetAddressNode;
+import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.nodes.type.StampTool;
 import jdk.graal.compiler.nodes.util.GraphUtil;
 import jdk.vm.ci.meta.JavaKind;
@@ -112,20 +111,17 @@ public class CardTableBarrierSet implements BarrierSet {
     }
 
     @Override
-    public void addBarriers(FixedAccessNode n) {
+    public void addBarriers(FixedAccessNode n, CoreProviders context) {
         if (n instanceof ReadNode) {
             // nothing to do
-        } else if (n instanceof WriteNode) {
-            WriteNode write = (WriteNode) n;
-            addWriteBarrier(write, write.value());
-        } else if (n instanceof LoweredAtomicReadAndWriteNode) {
-            LoweredAtomicReadAndWriteNode atomic = (LoweredAtomicReadAndWriteNode) n;
-            addWriteBarrier(atomic, atomic.getNewValue());
-        } else if (n instanceof AbstractCompareAndSwapNode) {
-            AbstractCompareAndSwapNode cmpSwap = (AbstractCompareAndSwapNode) n;
-            addWriteBarrier(cmpSwap, cmpSwap.getNewValue());
-        } else if (n instanceof ArrayRangeWrite) {
-            addArrayRangeBarriers((ArrayRangeWrite) n);
+        } else if (n instanceof WriteNode write) {
+            addWriteBarrier(write, write.value(), context);
+        } else if (n instanceof LoweredAtomicReadAndWriteNode atomic) {
+            addWriteBarrier(atomic, atomic.getNewValue(), context);
+        } else if (n instanceof AbstractCompareAndSwapNode cmpSwap) {
+            addWriteBarrier(cmpSwap, cmpSwap.getNewValue(), context);
+        } else if (n instanceof ArrayRangeWrite rangeWrite) {
+            addArrayRangeBarriers(rangeWrite, context);
         } else {
             GraalError.guarantee(n.getBarrierType() == BarrierType.NONE, "missed a node that requires a GC barrier: %s", n.getClass());
         }
@@ -175,25 +171,23 @@ public class CardTableBarrierSet implements BarrierSet {
         if (n instanceof ReadNode) {
             return false;
         } else if (n instanceof WriteNode || n instanceof LoweredAtomicReadAndWriteNode || n instanceof AbstractCompareAndSwapNode) {
-            return barrier instanceof SerialWriteBarrierNode && matches(n, (SerialWriteBarrierNode) barrier);
-        } else if (n instanceof ArrayRangeWrite) {
-            return barrier instanceof SerialArrayRangeWriteBarrierNode && matches((ArrayRangeWrite) n, (SerialArrayRangeWriteBarrierNode) barrier);
+            return barrier instanceof SerialWriteBarrierNode s && matches(n, s);
+        } else if (n instanceof ArrayRangeWrite w) {
+            return (barrier instanceof SerialArrayRangeWriteBarrierNode r && matches(w, r)) || (barrier instanceof SerialWriteBarrierNode s && !s.usePrecise() && matches(n, s));
         } else {
             throw GraalError.shouldNotReachHere("Unexpected node: " + n.getClass()); // ExcludeFromJacocoGeneratedReport
         }
     }
 
-    public void addArrayRangeBarriers(ArrayRangeWrite write) {
+    private void addArrayRangeBarriers(ArrayRangeWrite write, CoreProviders context) {
         if (arrayRangeWriteRequiresBarrier(write)) {
-            StructuredGraph graph = write.asNode().graph();
-            SerialArrayRangeWriteBarrierNode serialArrayRangeWriteBarrier = graph.add(new SerialArrayRangeWriteBarrierNode(write.getAddress(), write.getLength(), write.getElementStride()));
-            graph.addAfterFixed(write.postBarrierInsertionPosition(), serialArrayRangeWriteBarrier);
+            addSerialPostRangeWriteBarrier(write, context);
         }
     }
 
-    private void addWriteBarrier(FixedAccessNode node, ValueNode writtenValue) {
+    private void addWriteBarrier(FixedAccessNode node, ValueNode writtenValue, CoreProviders context) {
         if (needsWriteBarrier(node, writtenValue)) {
-            addSerialPostWriteBarrier(node, node.getAddress(), node.graph());
+            addSerialPostWriteBarrier(node, context);
         }
     }
 
@@ -228,15 +222,34 @@ public class CardTableBarrierSet implements BarrierSet {
     }
 
     private static boolean hasWriteBarrier(ArrayRangeWrite write) {
-        FixedWithNextNode node = write.asFixedWithNextNode();
-        return node.next() instanceof SerialArrayRangeWriteBarrierNode && matches(write, (SerialArrayRangeWriteBarrierNode) node.next());
+        FixedAccessNode node = (FixedAccessNode) write;
+        return (node.next() instanceof SerialArrayRangeWriteBarrierNode r && matches(write, r)) || (node.next() instanceof SerialWriteBarrierNode s && !s.usePrecise() && matches(node, s));
     }
 
-    private static void addSerialPostWriteBarrier(FixedAccessNode node, AddressNode address, StructuredGraph graph) {
-        // Use a precise barrier for everything that might be an array write. Being too precise with
-        // the barriers does not cause any correctness issues.
-        boolean precise = node.getBarrierType() != BarrierType.FIELD && node.getBarrierType() != BarrierType.AS_NO_KEEPALIVE_WRITE;
-        graph.addAfterFixed(node, graph.add(new SerialWriteBarrierNode(address, precise)));
+    private void addSerialPostRangeWriteBarrier(ArrayRangeWrite write, CoreProviders context) {
+        FixedAccessNode node = (FixedAccessNode) write;
+        StructuredGraph graph = node.graph();
+        if (!barrierPrecise(node, write.getAddress().getBase(), context)) {
+            SerialWriteBarrierNode barrier = graph.add(new SerialWriteBarrierNode(node.getAddress(), false));
+            graph.addAfterFixed(node, barrier);
+            return;
+        }
+
+        SerialArrayRangeWriteBarrierNode barrier = graph.add(new SerialArrayRangeWriteBarrierNode(write.getAddress(), write.getLength(), write.getElementStride()));
+        graph.addAfterFixed(write.postBarrierInsertionPosition(), barrier);
+    }
+
+    private void addSerialPostWriteBarrier(FixedAccessNode node, CoreProviders context) {
+        StructuredGraph graph = node.graph();
+        boolean precise = barrierPrecise(node, node.getAddress().getBase(), context);
+        graph.addAfterFixed(node, graph.add(new SerialWriteBarrierNode(node.getAddress(), precise)));
+    }
+
+    /**
+     * Decide if a precise barrier is needed for an access.
+     */
+    protected boolean barrierPrecise(FixedAccessNode node, @SuppressWarnings("unused") ValueNode base, @SuppressWarnings("unused") CoreProviders context) {
+        return node.getBarrierType() != BarrierType.FIELD && node.getBarrierType() != BarrierType.AS_NO_KEEPALIVE_WRITE;
     }
 
     private static boolean isNonNullObjectValue(ValueNode value) {