+++ substratevm/src/com.oracle.graal.pointsto/src/com/oracle/graal/pointsto/flow/StoreFieldTypeFlow.java
@@ -103,10 +103,10 @@ public abstract class StoreFieldTypeFlow extends AccessFieldTypeFlow {
     /**
      * The state of the StoreFieldTypeFlow reflects the state of the stored value. The
      * StoreFieldTypeFlow is an observer of the receiver flow, i.e. flow modeling the receiver
-     * object of the store operation..
+     * object of the store operation.
      *
      * Every time the state of the receiver flow changes the corresponding field flows are added as
-     * uses to the store field flow. Thus the stored value is propagated to the store field flow
+     * uses to the store field flow. Thus, the stored value is propagated to the store field flow
      * into the field flows.
      */
     public static class StoreInstanceFieldTypeFlow extends StoreFieldTypeFlow {